"Update formatting of product value in cart to include item quantity"
This commit updates the formatting of the product value in the cart to include the item quantity. Previously, the formatting only included the product's base value, but now it takes into account the quantity of the item in the cart. This results in a more accurate representation of the total cost of the item in the cart.