Add reset button to ProductForm component
The diff adds a reset button to the ProductForm component, which allows users to clear all form fields. The reset button is implemented using the IconButton component from MUI and the react-hook-form library's reset function. The button is styled using the warning color and displays a refresh icon. Additionally, a tooltip was added to provide a description of the button's functionality. The diff also includes some minor adjustments to the layout of the form fields.