"Added Badge component and IoIosCart icon to display cart item count in DefaultLayout"
This commit updates the DefaultLayout component to display the number of items in the cart using the Badge component from @mui/material and the IoIosCart icon from react-icons/io. The showBuildProductsSnackBar function is modified to not pass the autoHideDuration property explicitly, as it is already set to the default value of 3000. The previous implementation of displaying the cart item count using a Typography component is replaced with the new Badge component.